chore: fix ruff formatting and extract EnumDeliveryStatus to own file
Fixes Lint (ruff format) and Architecture validator (mixed types in
model_delivery_attempt.py). Enum now lives in enum_delivery_status.py.
