+++ src/omnibase_infra/services/retry_worker/models/__init__.py
@@ -2,8 +2,10 @@
 # Copyright (c) 2025 OmniNode Team
 """Models for the RetryWorker service."""
 
-from omnibase_infra.services.retry_worker.models.model_delivery_attempt import (
+from omnibase_infra.services.retry_worker.models.enum_delivery_status import (
     EnumDeliveryStatus,
+)
+from omnibase_infra.services.retry_worker.models.model_delivery_attempt import (
     ModelDeliveryAttempt,
 )
 from omnibase_infra.services.retry_worker.models.model_retry_result import (
